+++ src/kstd/basic.hh
@@ -43,8 +43,8 @@ struct Defer {
 // (as opposed to a compile-time N). Declared here rather than in
 // array.hh since string.hh needs it for a forward declaration
 // before array.hh itself is reachable (see string.hh comment).
-inline constexpr usize DYNAMIC_EXTENT = static_cast<usize>(-1);
+inline constexpr s64 DYNAMIC_EXTENT = -1;
 
 namespace mem {
-    inline constexpr usize PAGE_SIZE = 4096;
+    inline constexpr s64 PAGE_SIZE = 4096;
 }
