+++ src/kstd/ring_buffer.hh
@@ -9,12 +9,12 @@ struct Ring_Buffer {
     static_assert(N > 0);
 
     Static_Array<T, N> data;
-    usize              head = 0;
-    usize              tail = 0;
-    usize              size = 0;
+    s64                head = 0;
+    s64                tail = 0;
+    s64                size = 0;
 
     auto push_back(const T& element) -> void {
-        kstd_assert(size < N, "push_back on full Ring_Buffer");
+        kstd_assert(size < static_cast<s64>(N), "push_back on full Ring_Buffer");
 
         data[tail] = element;
         tail = (tail + 1) % N;
@@ -47,7 +47,7 @@ struct Ring_Buffer {
     auto format() const -> string {
         String_Builder builder;
         builder.append("[");
-        for (usize index = 0; index < size; ++index) {
+        for (s64 index = 0; index < size; ++index) {
             if (index > 0) builder.append(", ");
             builder.print(data[(head + index) % N]);
         }
